+++ src/renderer/components/layout/Sidebar.tsx
@@ -8,7 +8,7 @@ import ChevronLeftIcon from '@mui/icons-material/ChevronLeft';
 import ChevronRightIcon from '@mui/icons-material/ChevronRight';
 import HomeIcon from '@mui/icons-material/Home';
 import type { SvgIconComponent } from "@mui/icons-material";
-import { RoutePath } from "../constants/routes";
+import { RoutePath } from "../../constants/routes";
 
 // 导航项类型定义
 interface NavItem {